rename container/engine.rs -> shim/shim.rs
