+++ crates/containerd-shim-wasm/src/shim/log.rs
