+++ crates/containerd-shim-wasm/src/shim/path.rs
