+++ crates/containerd-shim-wasm/src/shim/wasm.rs
